requirements.txt
Flask
gunicorn
sklearn
pandas
numpy
seaborn
ipykernel
PyYAML